# src/generated/openapi/java/com/example/openapi/model/InsuranceUpdateRequest.java
/*
 * Persons API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.model;

import java.util.Objects;
import java.util.Arrays;
import java.util.Map;
import java.util.function.Function;
import java.util.stream.Collectors;
import java.time.LocalDate;
import com.fasterxml.jackson.annotation.*;
import io.micronaut.serde.annotation.Serdeable;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;
import io.micronaut.core.annotation.Nullable;
import jakarta.annotation.Generated;

/**
 * InsuranceUpdateRequest
 */
@JsonPropertyOrder({
    InsuranceUpdateRequest.JSON_PROPERTY_INSURER_NAME,
    InsuranceUpdateRequest.JSON_PROPERTY_PLAN,
    InsuranceUpdateRequest.JSON_PROPERTY_AMOUNT,
    InsuranceUpdateRequest.JSON_PROPERTY_EXPIRY_DATE,
})
@Serdeable
@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
public class InsuranceUpdateRequest {

    public static final String JSON_PROPERTY_INSURER_NAME = "insurerName";
    public static final String JSON_PROPERTY_PLAN = "plan";
    public static final String JSON_PROPERTY_AMOUNT = "amount";
    public static final String JSON_PROPERTY_EXPIRY_DATE = "expiryDate";

    @NotNull
    @JsonProperty(JSON_PROPERTY_INSURER_NAME)
    private String insurerName;

    @NotNull
    @JsonProperty(JSON_PROPERTY_PLAN)
    private String plan;

    @NotNull
    @JsonProperty(JSON_PROPERTY_AMOUNT)
    private Double amount;

    @NotNull
    @JsonProperty(JSON_PROPERTY_EXPIRY_DATE)
    private LocalDate expiryDate;

    public InsuranceUpdateRequest(String insurerName, String plan, Double amount, LocalDate expiryDate) {
        this.insurerName = insurerName;
        this.plan = plan;
        this.amount = amount;
        this.expiryDate = expiryDate;
    }

    /**
     * @return the insurerName property value
     */
    public String getInsurerName() {
        return insurerName;
    }

    /**
     * Set the insurerName property value
     *
     * @param insurerName property value to set
     */
    public void setInsurerName(String insurerName) {
        this.insurerName = insurerName;
    }

    /**
     * Set insurerName in a chainable fashion.
     *
     * @return The same instance of InsuranceUpdateRequest for chaining.
     */
    public InsuranceUpdateRequest insurerName(String insurerName) {
        this.insurerName = insurerName;
        return this;
    }

    /**
     * @return the plan property value
     */
    public String getPlan() {
        return plan;
    }

    /**
     * Set the plan property value
     *
     * @param plan property value to set
     */
    public void setPlan(String plan) {
        this.plan = plan;
    }

    /**
     * Set plan in a chainable fashion.
     *
     * @return The same instance of InsuranceUpdateRequest for chaining.
     */
    public InsuranceUpdateRequest plan(String plan) {
        this.plan = plan;
        return this;
    }

    /**
     * @return the amount property value
     */
    public Double getAmount() {
        return amount;
    }

    /**
     * Set the amount property value
     *
     * @param amount property value to set
     */
    public void setAmount(Double amount) {
        this.amount = amount;
    }

    /**
     * Set amount in a chainable fashion.
     *
     * @return The same instance of InsuranceUpdateRequest for chaining.
     */
    public InsuranceUpdateRequest amount(Double amount) {
        this.amount = amount;
        return this;
    }

    /**
     * @return the expiryDate property value
     */
    public LocalDate getExpiryDate() {
        return expiryDate;
    }

    /**
     * Set the expiryDate property value
     *
     * @param expiryDate property value to set
     */
    public void setExpiryDate(LocalDate expiryDate) {
        this.expiryDate = expiryDate;
    }

    /**
     * Set expiryDate in a chainable fashion.
     *
     * @return The same instance of InsuranceUpdateRequest for chaining.
     */
    public InsuranceUpdateRequest expiryDate(LocalDate expiryDate) {
        this.expiryDate = expiryDate;
        return this;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        var insuranceUpdateRequest = (InsuranceUpdateRequest) o;
        return Objects.equals(this.insurerName, insuranceUpdateRequest.insurerName)
            && Objects.equals(this.plan, insuranceUpdateRequest.plan)
            && Objects.equals(this.amount, insuranceUpdateRequest.amount)
            && Objects.equals(this.expiryDate, insuranceUpdateRequest.expiryDate);
    }

    @Override
    public int hashCode() {
        return Objects.hash(insurerName, plan, amount, expiryDate);
    }

    @Override
    public String toString() {
        return "InsuranceUpdateRequest("
            + "insurerName: " + getInsurerName() + ", "
            + "plan: " + getPlan() + ", "
            + "amount: " + getAmount() + ", "
            + "expiryDate: " + getExpiryDate()
            + ")";
    }

}