# src/generated/openapi/java/com/example/openapi/model/InsurancesResponse.java
/*
 * Persons API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.model;

import java.util.Objects;
import java.util.Arrays;
import java.util.Map;
import java.util.function.Function;
import java.util.stream.Collectors;
import com.example.openapi.model.Insurance;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import com.fasterxml.jackson.annotation.*;
import io.micronaut.serde.annotation.Serdeable;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;
import io.micronaut.core.annotation.Nullable;
import jakarta.annotation.Generated;

/**
 * InsurancesResponse
 */
@JsonPropertyOrder({
    InsurancesResponse.JSON_PROPERTY_CONTENT,
    InsurancesResponse.JSON_PROPERTY_TOTAL_SIZE,
    InsurancesResponse.JSON_PROPERTY_TOTAL_PAGES,
})
@Serdeable
@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
public class InsurancesResponse {

    public static final String JSON_PROPERTY_CONTENT = "content";
    public static final String JSON_PROPERTY_TOTAL_SIZE = "totalSize";
    public static final String JSON_PROPERTY_TOTAL_PAGES = "totalPages";

    @Nullable(inherited = true)
    @JsonProperty(JSON_PROPERTY_CONTENT)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private List<@Valid Insurance> content;

    @Nullable(inherited = true)
    @JsonProperty(JSON_PROPERTY_TOTAL_SIZE)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private Integer totalSize;

    @Nullable(inherited = true)
    @JsonProperty(JSON_PROPERTY_TOTAL_PAGES)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private Integer totalPages;

    /**
     * @return the content property value
     */
    public List<@Valid Insurance> getContent() {
        return content;
    }

    /**
     * Set the content property value
     *
     * @param content property value to set
     */
    public void setContent(List<@Valid Insurance> content) {
        this.content = content;
    }

    /**
     * Set content in a chainable fashion.
     *
     * @return The same instance of InsurancesResponse for chaining.
     */
    public InsurancesResponse content(List<@Valid Insurance> content) {
        this.content = content;
        return this;
    }

    /**
     * Add an item to the content property in a chainable fashion.
     *
     * @return The same instance of InsurancesResponse for chaining.
     */
    public InsurancesResponse addContentItem(Insurance contentItem) {
        if (content == null) {
            content = new ArrayList<>();
        }
        content.add(contentItem);
        return this;
    }

    /**
     * @return the totalSize property value
     */
    public Integer getTotalSize() {
        return totalSize;
    }

    /**
     * Set the totalSize property value
     *
     * @param totalSize property value to set
     */
    public void setTotalSize(Integer totalSize) {
        this.totalSize = totalSize;
    }

    /**
     * Set totalSize in a chainable fashion.
     *
     * @return The same instance of InsurancesResponse for chaining.
     */
    public InsurancesResponse totalSize(Integer totalSize) {
        this.totalSize = totalSize;
        return this;
    }

    /**
     * @return the totalPages property value
     */
    public Integer getTotalPages() {
        return totalPages;
    }

    /**
     * Set the totalPages property value
     *
     * @param totalPages property value to set
     */
    public void setTotalPages(Integer totalPages) {
        this.totalPages = totalPages;
    }

    /**
     * Set totalPages in a chainable fashion.
     *
     * @return The same instance of InsurancesResponse for chaining.
     */
    public InsurancesResponse totalPages(Integer totalPages) {
        this.totalPages = totalPages;
        return this;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        var insurancesResponse = (InsurancesResponse) o;
        return Objects.equals(this.content, insurancesResponse.content)
            && Objects.equals(this.totalSize, insurancesResponse.totalSize)
            && Objects.equals(this.totalPages, insurancesResponse.totalPages);
    }

    @Override
    public int hashCode() {
        return Objects.hash(content, totalSize, totalPages);
    }

    @Override
    public String toString() {
        return "InsurancesResponse("
            + "content: " + getContent() + ", "
            + "totalSize: " + getTotalSize() + ", "
            + "totalPages: " + getTotalPages()
            + ")";
    }

}